tsconfig + cleanup
