+++ native-new/Cargo.toml
@@ -21,3 +21,4 @@ gleam = "0.6.9"
 ordered-float = "1.0.1"
 log = "0.4.6"
 env_logger = "0.6.0"
+font-loader = "0.8.0"
