+++ package.json
@@ -31,6 +31,7 @@
   },
   "devDependencies": {
     "@types/node": "^10.12.1",
+    "@types/react": "^16.8.6",
     "@types/react-native": "^0.57.18",
     "@types/yoga-layout": "^1.9.0",
     "react": "^16.8.3",
