--- tscodegen/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "tscodegen"
-version = "0.1.0"
-authors = ["Kamil Tomšík <info@tomsik.cz>"]
-edition = "2018"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = { version = "0.15.26", features = ["extra-traits"] }
-quote = "0.6.11"
